refactor(fe): make bar components
